+++ backend/recipe_api/serializers.py
@@ -185,8 +185,8 @@ class RecipeSerializer(serializers.ModelSerializer):
 
 class RecipeCreateIngridientSerializer(serializers.ModelSerializer):
     id = serializers.PrimaryKeyRelatedField(queryset=Ingredient.objects.all())
-    amount = serializers.IntegerField()
-
+    # amount = serializers.IntegerField()
+    amount = serializers.FloatField()
     class Meta:
         model = RecipeIngredient
         fields = (
